Address review comments, upgrade probe.gl, polish table logging
