+++ src/utils/index.js
@@ -1,7 +1,8 @@
 export {isBrowser} from './is-browser';
 export {promisify} from './promise-utils';
 export {uid, isPowerOfTwo, isObjectEmpty} from './utils';
-export {default as log, formatValue} from './log';
+export {default as log} from './log';
+export {default as oldLog, formatValue} from './old-log';
 export {
   getGLTypeFromTypedArray, getTypedArrayFromGLType, flipRows, scalePixels
 } from './typed-array-utils';
